Changed function declarations
guassianElim() now takes in the index vector by reference, so it can be used with the matrix for the solve function

solve() added and takes in matrix, index vector, and b vector representing dependent variable values for the system of equations. This allows for multiple b vectors to be used for a single system of equations.