Modified gaussElim function, added solve function, and moved printMat to main.
Solve is used with after a system of equations has been processed by the gaussElim function. It takes in the index vector l by reference, as well as a vector b which is the dependent variables for each equation in the system. This allows for multiple b vectors to be used for one system of equations. Solve returns a vector of each x value from x_1 ... x_n.